requirements.txt
certifi==2019.9.11
chardet==3.0.4
gspread==3.1.0
httplib2==0.14.0
idna==2.8
oauth2client==4.1.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.0
requests==2.22.0
rsa==4.0
six==1.12.0
slackbot==0.5.4
slacker==0.13.0
urllib3==1.25.6
websocket-client==0.44.0
